pyproject.toml
[tool.poetry]
name = "simple-pastebin"
version = "0.1.0"
description = ""
authors = ["korruz <korruz@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
aiosqlite = "^0.20.0"
flask = {extras = ["async"], version = "^3.0.3"}
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
pygments = "^2.18.0"
apscheduler = "^3.10.4"
greenlet = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
